Dictamenes para Crédito de Sustentabilidad y Modificaciones al Crédito Terminados.

+++ Entidades/MCDictamenTecnico.cs
@@ -12,7 +12,7 @@ namespace Entidades
         public int MCDT_IDModificacionesCredito { get; set; }
         public int MCDT_Procedencia { get; set; }
         public string MCDT_MotivosProcedencia { get; set; }
-        public float MCDT_MontoSugerido { get; set; }
+        public double MCDT_MontoSugerido { get; set; }
         public DateTime MCDT_FechaDictaminacion { get; set; }
         public string MCDT_NoAsesorTecnico { get; set; }
         public string MCDT_UsuarioDominio { get; set; }
